third_party/blink/web_tests/http/tests/asmjs/asm-warnings-expected.txt

CONSOLE WARNING: Invalid asm.js: Bad variable declaration
CONSOLE WARNING: Invalid asm.js: Can only use immutable variables in global definition
